requirements.txt
# Core data + ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
prophet>=1.1.5
statsmodels>=0.14.0
linearmodels==4.31
rdrobust>=1.0.0

# Snowflake
snowflake-connector-python>=3.5.0
snowflake-snowpark-python>=1.11.0

# dbt
dbt-core>=1.7.0
dbt-snowflake>=1.7.0

# Airflow (install separately via pip or Docker)
apache-airflow>=2.8.0

# API / HTTP
requests>=2.31.0
httpx>=0.26.0
tenacity>=8.2.3

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0
altair>=5.2.0

# Env / config
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
responses>=0.24.1

# Formatting / linting
black>=23.0.0
ruff>=0.1.0

# Utilities
loguru>=0.7.0
pyarrow>=14.0.0
